requirements.txt
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fonttools==4.28.4
idna==3.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.18.1
jsonschema==4.3.0
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.21.4
packaging==21.3
pandas==1.3.5
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pyarrow==6.0.1
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
scipy==1.7.3
six==1.16.0
statsmodels==0.13.1
tenacity==8.0.1
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
zipp==3.6.0
